+++ trunk/src/ta3d/src/ingame/sidedata.h
@@ -28,8 +28,8 @@ namespace TA3D
 
 	struct IntrElementCoords // INT_ELEMENT
 	{
-		uint16	x1, y1;
-		uint16	x2, y2;
+		int	x1, y1;
+		int	x2, y2;
 
 	}; // IntrElementCoords
